move deciamlavg
