+++ datafusion/physical-expr/src/aggregate/mod.rs
@@ -54,7 +54,12 @@ pub(crate) mod variance;
 
 pub mod build_in;
 pub mod moving_min_max;
-pub mod utils;
+pub mod utils {
+    pub use datafusion_physical_expr_common::aggregate::utils::{
+        adjust_output_array, down_cast_any_ref, get_accum_scalar_values_as_arrays,
+        get_sort_options, ordering_fields, DecimalAverager, Hashable,
+    };
+}
 
 /// Checks whether the given aggregate expression is order-sensitive.
 /// For instance, a `SUM` aggregation doesn't depend on the order of its inputs.
