--- datafusion/physical-expr/src/aggregate/utils.rs
@@ -1,110 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-//! Utilities used in aggregates
-
-// For backwards compatibility
-pub use datafusion_physical_expr_common::aggregate::utils::{
-    adjust_output_array, down_cast_any_ref, get_accum_scalar_values_as_arrays,
-    get_sort_options, ordering_fields, Hashable,
-};
-
-use arrow::array::ArrowNativeTypeOp;
-use arrow_array::types::DecimalType;
-use arrow_buffer::ArrowNativeType;
-use datafusion_common::{exec_err, DataFusionError, Result};
-
-// TODO: Move to functions-aggregate crate
-/// Computes averages for `Decimal128`/`Decimal256` values, checking for overflow
-///
-/// This is needed because different precisions for Decimal128/Decimal256 can
-/// store different ranges of values and thus sum/count may not fit in
-/// the target type.
-///
-/// For example, the precision is 3, the max of value is `999` and the min
-/// value is `-999`
-pub(crate) struct DecimalAverager<T: DecimalType> {
-    /// scale factor for sum values (10^sum_scale)
-    sum_mul: T::Native,
-    /// scale factor for target (10^target_scale)
-    target_mul: T::Native,
-    /// the output precision
-    target_precision: u8,
-}
-
-impl<T: DecimalType> DecimalAverager<T> {
-    /// Create a new `DecimalAverager`:
-    ///
-    /// * sum_scale: the scale of `sum` values passed to [`Self::avg`]
-    /// * target_precision: the output precision
-    /// * target_scale: the output scale
-    ///
-    /// Errors if the resulting data can not be stored
-    pub fn try_new(
-        sum_scale: i8,
-        target_precision: u8,
-        target_scale: i8,
-    ) -> Result<Self> {
-        let sum_mul = T::Native::from_usize(10_usize)
-            .map(|b| b.pow_wrapping(sum_scale as u32))
-            .ok_or(DataFusionError::Internal(
-                "Failed to compute sum_mul in DecimalAverager".to_string(),
-            ))?;
-
-        let target_mul = T::Native::from_usize(10_usize)
-            .map(|b| b.pow_wrapping(target_scale as u32))
-            .ok_or(DataFusionError::Internal(
-                "Failed to compute target_mul in DecimalAverager".to_string(),
-            ))?;
-
-        if target_mul >= sum_mul {
-            Ok(Self {
-                sum_mul,
-                target_mul,
-                target_precision,
-            })
-        } else {
-            // can't convert the lit decimal to the returned data type
-            exec_err!("Arithmetic Overflow in AvgAccumulator")
-        }
-    }
-
-    /// Returns the `sum`/`count` as a i128/i256 Decimal128/Decimal256 with
-    /// target_scale and target_precision and reporting overflow.
-    ///
-    /// * sum: The total sum value stored as Decimal128 with sum_scale
-    /// (passed to `Self::try_new`)
-    /// * count: total count, stored as a i128/i256 (*NOT* a Decimal128/Decimal256 value)
-    #[inline(always)]
-    pub fn avg(&self, sum: T::Native, count: T::Native) -> Result<T::Native> {
-        if let Ok(value) = sum.mul_checked(self.target_mul.div_wrapping(self.sum_mul)) {
-            let new_value = value.div_wrapping(count);
-
-            let validate =
-                T::validate_decimal_precision(new_value, self.target_precision);
-
-            if validate.is_ok() {
-                Ok(new_value)
-            } else {
-                exec_err!("Arithmetic Overflow in AvgAccumulator")
-            }
-        } else {
-            // can't convert the lit decimal to the returned data type
-            exec_err!("Arithmetic Overflow in AvgAccumulator")
-        }
-    }
-}
